Move nvtx_utils.hpp to public headers
