+++ cpp/src/binaryop/compiled/launcher.cuh
@@ -17,7 +17,7 @@
 #ifndef COMPILED_BINARY_OPS_LAUNCHER_H
 #define COMPILED_BINARY_OPS_LAUNCHER_H
 
-#include <utilities/nvtx/nvtx_utils.hpp>
+#include <cudf/utilities/nvtx_utils.hpp>
 #include <utilities/error_utils.hpp>
 #include <cudf/cudf.h>
