+++ cpp/src/hash/legacy/hashing.cu
@@ -23,7 +23,7 @@
 #include <table/legacy/device_table.cuh>
 #include "hash/hash_functions.cuh"
 #include <utilities/int_fastdiv.h>
-#include <utilities/nvtx/nvtx_utils.hpp>
+#include <cudf/utilities/nvtx_utils.hpp>
 #include <copying/legacy/scatter.hpp>
 #include <cudf/types.hpp>
